pyproject.toml
[project]
name = "context-pilot"
version = "0.1.0"
description = "Context Pilot Tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ADK Framework
    "google-adk>=1.20.0",
    
    # Web Server (ADK needs these)
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    
    # Essential utilities (needed by our code)
    "httpx>=0.28.1",  # Required by GeminiRawEmbedding
    "python-dotenv>=1.0.1",
    
    # RAG Stack (minimal)
    "llama-index-core>=0.10.0",  # Core only, not full llama-index
    "llama-index-llms-gemini>=0.1.0",
    "llama-index-embeddings-gemini>=0.1.0",
]

[project.scripts]
context-pilot = "context_pilot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["context_pilot"]

[tool.hatch.build.targets.sdist]
include = [
    "context_pilot",
    "pyproject.toml",
    "README.md"
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
